+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### [Unreleased]
 ## [0.16.8] - 2020-04-14
+- Changed activity availability
 - Small fix
 
 ## [0.16.7] - 2020-04-10